Increase test cover (#69)
* Increase WhereByRequest function

In `WhereByRequest` function it is only allowed to search for simple `WHERE` and
for fields type JSONb. Searches for JSON-type fields are not yet
allowed. Rows removed in function never came into use with what we
support today.

* Improvements tests in Query function

The code removed in Query function isn't necessary becuase
chkIndentifer is valid only to small sql strings.

* Improvements PaginateIfPossible function

* Improvements in Insert function

* Improvements in Delete function

* Improvements in Update function

* Improvements in GetDatabases function

* Improvements in GetSchemas function

* Improvements in OrderByRequest function

Done otherwise, `reqOrder` variable returned as array of strings.
Now, this var comes as string and can be easily checked if during the
request came empty or not.

* Rewrite test execution in databases_test

* Improvements functions based on 65ba76

* Add more tests to GetTables function

* Add more tests to GetTablesByDatabaseAndSchema

* Add more tests to SelectFromTables function

* Add more tests to InsertInTables function

* Add more tests to DeleteFromTables function

* Add more tests to UpdateTable function

* Change expectation in DeleteFromTables test

* Fix tests running with invalid function

* Add check sql injection in OrderByRequest

* Remove unneccessary check to path vars

By default, the lack of this vars is handled by gorilla mux,
returning to user http status 3O1.

* Improvements JoinByRequest function

Add checkInvalidIdentifier...

* Add check invalid identifier in other functions

The function PaginateIfPossible isn't neccessary the addition, because
in querystrings _page and _page_size ever is int values.

* Improve tests in controllers