Add examples to the Filter section of the README
Not knowing a lot about the project, I've had to skimm through the
source code to understand how the different filter operators could be
used.

This commit formats the table containing the filtering options and adds
examples on how to use each filter and also how they are translated into
an SQL statement.

Signed-off-by: Diogo Osório <diogo.g.osorio@gmail.com>